requirements.txt
requests>=2.31.0
python-dateutil>=2.9.0